Moved continuous output tests to test_data
